Codechange #8088: Number of towns silently limited groundwork, provide the framework to extend original town name generators enabling previous restrictions to be overcome.
