+++ src/strings.cpp
@@ -347,7 +347,11 @@ void GetStringWithArgs(StringBuilder &builder, StringID string, StringParameters
 		case TEXT_TAB_TOWN:
 			if (IsInsideMM(string, SPECSTR_TOWNNAME_START, SPECSTR_TOWNNAME_END) && !game_script) {
 				try {
-					GenerateTownNameString(builder, string - SPECSTR_TOWNNAME_START, args.GetNextParameter<uint32_t>());
+					bool use_original_generator;
+					use_original_generator = *std::get_if<uint64_t>(&args.GetParam((size_t)1));
+					uint32_t randseed;
+					randseed = *std::get_if<uint64_t>(&args.GetParam((size_t)0));
+					GenerateTownNameString(builder, string - SPECSTR_TOWNNAME_START, randseed, use_original_generator);
 				} catch (const std::runtime_error &e) {
 					Debug(misc, 0, "GetStringWithArgs: {}", e.what());
 					builder += "(invalid string parameter)";
@@ -2008,7 +2012,7 @@ static bool GetSpecialNameString(StringBuilder &builder, StringID string, String
 
 	/* TownName Transport company names, with the appropriate town name. */
 	if (IsInsideMM(string, SPECSTR_COMPANY_NAME_START, SPECSTR_COMPANY_NAME_END)) {
-		GenerateTownNameString(builder, string - SPECSTR_COMPANY_NAME_START, args.GetNextParameter<uint32_t>());
+		GenerateTownNameString(builder, string - SPECSTR_COMPANY_NAME_START, args.GetNextParameter<uint32_t>(), true);
 		builder += " Transport";
 		return true;
 	}
