+++ src/table/townname_extended.h
@@ -1,0 +1,89 @@
+/*
+ * This file is part of OpenTTD.
+ * OpenTTD is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, version 2.
+ * OpenTTD is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with OpenTTD. If not, see <https://www.gnu.org/licenses/old-licenses/gpl-2.0>.
+ */
+
+/** @file townname_extended.h Namepart tables for the town name generator. */
+
+#include "../core/enum_type.hpp"
+
+static const std::string_view _name_french_prefix[] = {
+	"St.",
+	"Le",
+	"La"
+};
+
+static const std::string_view _name_french_semireal[] = {
+	"Agincourt",
+	"Lille",
+	"Dinan",
+	"Aubusson",
+	"Rodez",
+	"Bergerac",
+	"Bordeaux",
+	"Bayonne",
+	"Montpellier",
+	"Mont\u00e9limar",
+	"Valence",
+	"Digne",
+	"Nice",
+	"Cannes",
+	"Tropez",
+	"Marseille",
+	"Narbonne",
+	"S\u00e8te",
+	"Aurillac",
+	"Gu\u00e9ret",
+	"Creusot",
+	"Nevers",
+	"Auxerre",
+	"Versailles",
+	"Meaux",
+	"Ch\u00e2lons",
+	"Compi\u00e8gne",
+	"Metz",
+	"Chaumont",
+	"Langres",
+	"Bourg",
+	"Lyon",
+	"Vienne",
+	"Grenoble",
+	"Toulon",
+	"Rennes",
+	"Mans",
+	"Angers",
+	"Nantes",
+	"Ch\u00e2teauroux",
+	"Orl\u00e9ans",
+	"Lisieux",
+	"Cherbourg",
+	"Morlaix",
+	"Cognac",
+	"Agen",
+	"Tulle",
+	"Blois",
+	"Troyes",
+	"Charolles",
+	"Toulouse",
+	"Chamb\u00e9ry",
+	"Tours",
+	"Brieuc",
+	"Malo",
+	"Rochelle",
+	"Flour",
+	"Puy",
+	"Vichy",
+	"Valery",
+	"Beaujolais",
+	"Reims",
+	"Albi",
+	"Paris",
+	"Biarritz",
+	"B\u00e9ziers",
+	"N\u00eemes",
+	"Chamonix",
+	"Angoul\u00e8me",
+	"Alen\u00e7on",
+};
