+++ src/table/townname_original.h
@@ -5,7 +5,7 @@
  * See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with OpenTTD. If not, see <https://www.gnu.org/licenses/old-licenses/gpl-2.0>.
  */
 
-/** @file townname.h Namepart tables for the town name generator. */
+/** @file townname_original.h Namepart tables for the town name generator. */
 
 #include "../core/enum_type.hpp"
