+++ src/townname_extended.cpp
@@ -1,0 +1,186 @@
+/*
+ * This file is part of OpenTTD.
+ * OpenTTD is free software; you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, version 2.
+ * OpenTTD is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU General Public License for more details. You should have received a copy of the GNU General Public License along with OpenTTD. If not, see <https://www.gnu.org/licenses/old-licenses/gpl-2.0>.
+ */
+
+/** @file townname_extended.cpp %Town name generators.
+ *
+ * Original generators in townname_original.cpp must be maintained to return the same values for a given seed so that historic save games will load the same names as before.
+ * If the original generator fails to provide suitable names due to error, limit in available names, etc, it can be extended either with additional logic or completely new logic in a new function in this file.
+ */
+
+#include "stdafx.h"
+#include "string_func.h"
+#include "townname_type.h"
+#include "town.h"
+#include "strings_func.h"
+#include "core/random_func.hpp"
+#include "genworld.h"
+#include "gfx_layout.h"
+#include "strings_internal.h"
+
+#include "table/townname_extended.h"
+
+#include "safeguards.h"
+
+
+/**
+ * Generates a number from given seed.
+ * @param shift_by number of bits seed is shifted to the right
+ * @param max generated number is in interval 0...max-1
+ * @param seed seed
+ * @return seed transformed to a number from given range
+ */
+static inline uint32_t SeedChance(uint8_t shift_by, size_t max, uint32_t seed)
+{
+	return (GB(seed, shift_by, 16) * ClampTo<uint16_t>(max)) >> 16;
+}
+
+
+/**
+ * Generates a number from given seed. Uses different algorithm than SeedChance().
+ * @param shift_by number of bits seed is shifted to the right
+ * @param max generated number is in interval 0...max-1
+ * @param seed seed
+ * @return seed transformed to a number from given range
+ */
+static inline uint32_t SeedModChance(uint8_t shift_by, size_t max, uint32_t seed)
+{
+	/* This actually gives *MUCH* more even distribution of the values
+	 * than SeedChance(), which is absolutely horrible in that. If
+	 * you do not believe me, try with i.e. the Czech town names,
+	 * compare the words (nicely visible on prefixes) generated by
+	 * SeedChance() and SeedModChance(). Do not get discouraged by the
+	 * never-use-modulo myths, which hold true only for the linear
+	 * congruential generators (and Random() isn't such a generator).
+	 * --pasky
+	 * TODO: Perhaps we should use it for all the name generators? --pasky */
+	return (seed >> shift_by) % max;
+}
+
+
+/**
+ * Generates a number from given seed.
+ * @param shift_by number of bits seed is shifted to the right
+ * @param max generated number is in interval -bias...max-1
+ * @param seed seed
+ * @param bias minimum value that can be returned
+ * @return seed transformed to a number from given range
+ */
+static inline int32_t SeedChanceBias(uint8_t shift_by, size_t max, uint32_t seed, int bias)
+{
+	return SeedChance(shift_by, max + bias, seed) - bias;
+}
+
+
+/**
+ * Replaces a string beginning in 'org' with 'rep'.
+ * @param org     string to replace
+ * @param rep     string to be replaced with
+ * @param str     string of the town name
+ * @param start   the start index within the string for the town name
+ */
+static void ReplaceWords(std::string_view org, std::string_view rep, std::string &str, size_t start)
+{
+	if (str.compare(start, org.size(), org) == 0) str.replace(start, org.size(), rep);
+}
+
+
+/**
+ * Replaces english curses and ugly letter combinations by nicer ones.
+ * @param str      The string with the town name
+ * @param start    The start index into the string for the first town name
+ * @param original English (Original) generator was used
+ */
+static void ReplaceEnglishWords(std::string &str, size_t start, bool original)
+{
+	if (original) ReplaceWords("Ce", "Ke", str, start);
+	if (original) ReplaceWords("Ci", "Ki", str, start);
+	ReplaceWords("Cunt", "East", str, start);
+	ReplaceWords("Slag", "Pits", str, start);
+	ReplaceWords("Slut", "Edin", str, start);
+	if (!original) ReplaceWords("Fart", "Boot", str, start); // never happens with 'English (Original)'
+	ReplaceWords("Drar", "Quar", str, start);
+	ReplaceWords("Dreh", "Bash", str, start);
+	ReplaceWords("Frar", "Shor", str, start);
+	ReplaceWords("Grar", "Aber", str, start);
+	ReplaceWords("Brar", "Over", str, start);
+	ReplaceWords("Wrar", original ? "Inve" : "Stan", str, start);
+}
+
+
+/**
+ * Generates French town name from given seed.
+ * @param builder string builder
+ * @param seed town name seed
+ */
+static void MakeFrenchTownName(StringBuilder &builder, uint32_t seed)
+{
+	// ~50% of the time add a random prefix
+	if (seed % 2 == 0) {
+		builder += _name_french_prefix[seed % std::size(_name_french_prefix)];
+		builder += " ";
+	}
+	builder += _name_french_semireal[SeedChance(0, std::size(_name_french_semireal), seed)];
+}
+
+
+/**
+ *  Store the lang value so that we can defer to the original generators when an enhanced generator doesn't exist
+ */
+static size_t _lang;
+
+
+/**
+ * Generates original town name from lang and given seed.
+ * @param builder string builder
+ * @param seed town name seed
+ */
+static void MakeOriginalTownName(StringBuilder &builder, uint32_t seed)
+{
+	GenerateOriginalTownNameString(builder, _lang, seed);
+}
+
+/**
+ * Town name generators
+ * Order is important and must match original generators, when developing an extended generator create the function in this file and add it's name in the corect location below
+ */
+static TownNameGenerator *const _town_name_generators[] = {
+	MakeOriginalTownName,   // MakeEnglishOriginalTownName
+	MakeFrenchTownName,     // MakeFrenchTownName
+	MakeOriginalTownName,   // MakeGermanTownName
+	MakeOriginalTownName,   // MakeEnglishAdditionalTownName
+	MakeOriginalTownName,   // MakeSpanishTownName
+	MakeOriginalTownName,   // MakeSillyTownName
+	MakeOriginalTownName,   // MakeSwedishTownName
+	MakeOriginalTownName,   // MakeDutchTownName
+	MakeOriginalTownName,   // MakeFinnishTownName
+	MakeOriginalTownName,   // MakePolishTownName
+	MakeOriginalTownName,   // MakeSlovakTownName
+	MakeOriginalTownName,   // MakeNorwegianTownName
+	MakeOriginalTownName,   // MakeHungarianTownName
+	MakeOriginalTownName,   // MakeAustrianTownName
+	MakeOriginalTownName,   // MakeRomanianTownName
+	MakeOriginalTownName,   // MakeCzechTownName
+	MakeOriginalTownName,   // MakeSwissTownName
+	MakeOriginalTownName,   // MakeDanishTownName
+	MakeOriginalTownName,   // MakeTurkishTownName
+	MakeOriginalTownName,   // MakeItalianTownName
+	MakeOriginalTownName,   // MakeCatalanTownName
+};
+
+
+/**
+ * Generates town name from given seed.
+ * @param builder string builder to write to
+ * @param lang    town name language
+ * @param seed    generation seed
+ */
+void GenerateExtendedTownNameString(StringBuilder &builder, size_t lang, uint32_t seed)
+{
+	assert(lang < std::size(_town_name_generators));
+	_lang = lang;
+	return _town_name_generators[lang](builder, seed);
+}
